Extract as a generic helper
Can be used in other sidekiq-dependent RES gems.
